refactor: change file structure about user api dto
