+++ src/user/types/idpUserInfo.type.ts
@@ -1,0 +1,6 @@
+export type IdpUserInfo = {
+  sub: string;
+  name: string;
+  email: string;
+  student_id: string;
+};
